requirements.txt
Authlib==1.2.1
base36==0.1.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
Flask==3.0.0
Flask-Login @ git+https://github.com/maxcountryman/flask-login.git@7d98a49
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
python-gitlab==3.15.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
simpleeval==0.9.13
python-slugify==8.0.1
text-unidecode==1.3
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
WTForms==3.1.0
